// frontend/plugins/welcome/src/api/models/EntEquipment.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SUT SA Example API Patient
 * This is a sample server for SUT SE 2563
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    EntEquipmentEdges,
    EntEquipmentEdgesFromJSON,
    EntEquipmentEdgesFromJSONTyped,
    EntEquipmentEdgesToJSON,
} from './';

/**
 * 
 * @export
 * @interface EntEquipment
 */
export interface EntEquipment {
    /**
     * EQUIPMENTAMOUNT holds the value of the "EQUIPMENTAMOUNT" field.
     * @type {number}
     * @memberof EntEquipment
     */
    eQUIPMENTAMOUNT?: number;
    /**
     * EQUIPMENTDATE holds the value of the "EQUIPMENTDATE" field.
     * @type {string}
     * @memberof EntEquipment
     */
    eQUIPMENTDATE?: string;
    /**
     * EQUIPMENTDETAIL holds the value of the "EQUIPMENTDETAIL" field.
     * @type {string}
     * @memberof EntEquipment
     */
    eQUIPMENTDETAIL?: string;
    /**
     * EQUIPMENTNAME holds the value of the "EQUIPMENTNAME" field.
     * @type {string}
     * @memberof EntEquipment
     */
    eQUIPMENTNAME?: string;
    /**
     * 
     * @type {EntEquipmentEdges}
     * @memberof EntEquipment
     */
    edges?: EntEquipmentEdges;
    /**
     * ID of the ent.
     * @type {number}
     * @memberof EntEquipment
     */
    id?: number;
}

export function EntEquipmentFromJSON(json: any): EntEquipment {
    return EntEquipmentFromJSONTyped(json, false);
}

export function EntEquipmentFromJSONTyped(json: any, ignoreDiscriminator: boolean): EntEquipment {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'eQUIPMENTAMOUNT': !exists(json, 'EQUIPMENTAMOUNT') ? undefined : json['EQUIPMENTAMOUNT'],
        'eQUIPMENTDATE': !exists(json, 'EQUIPMENTDATE') ? undefined : json['EQUIPMENTDATE'],
        'eQUIPMENTDETAIL': !exists(json, 'EQUIPMENTDETAIL') ? undefined : json['EQUIPMENTDETAIL'],
        'eQUIPMENTNAME': !exists(json, 'EQUIPMENTNAME') ? undefined : json['EQUIPMENTNAME'],
        'edges': !exists(json, 'edges') ? undefined : EntEquipmentEdgesFromJSON(json['edges']),
        'id': !exists(json, 'id') ? undefined : json['id'],
    };
}

export function EntEquipmentToJSON(value?: EntEquipment | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'EQUIPMENTAMOUNT': value.eQUIPMENTAMOUNT,
        'EQUIPMENTDATE': value.eQUIPMENTDATE,
        'EQUIPMENTDETAIL': value.eQUIPMENTDETAIL,
        'EQUIPMENTNAME': value.eQUIPMENTNAME,
        'edges': EntEquipmentEdgesToJSON(value.edges),
        'id': value.id,
    };
}